+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog for aiodogstatsd
 
 ## 0.9.0 (2019-XX-XX)
-- Added timer contextmanager for easily timing events.
-- Added sample rate as class attribute, for setting sample rate class-wide.
+
+- Added sample rate as class attribute, for setting sample rate class-wide, #11 by @aviramha
+- Added timer contextmanager for easily timing events, #12 by @aviramha
+- Added Python 3.8.* support, #7
 
 ## 0.8.0 (2019-11-03)